Refactored gtm code to its own file
